fluxos definidos

+++ src/main/java/com/generation/blog/controller/TemaController.java
@@ -1,6 +1,9 @@
 package com.generation.blog.controller;
 
 import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -13,7 +16,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.generation.blog.model.Tema;
 import com.generation.blog.repository.TemaRepository;
@@ -38,24 +43,29 @@ public class TemaController {
 				.orElse(ResponseEntity.notFound().build());
 	}
 	
-	@GetMapping("/nome/{nome}")
+	@GetMapping("/nome/{nome}") // "descricao"
 	public ResponseEntity<List<Tema>> getByName(@PathVariable String nome){
 		return ResponseEntity.ok(repository.findAllByDescricaoContainingIgnoreCase(nome));
 	}
 	
 	@PostMapping
-	public ResponseEntity<Tema> post (@RequestBody Tema tema){
+	public ResponseEntity<Tema> post (@Valid @RequestBody Tema tema){
 		return ResponseEntity.status(HttpStatus.CREATED)
 				.body(repository.save(tema));
 	}
 	
-	@PutMapping
-	public ResponseEntity<Tema> put (@RequestBody Tema tema){
+	@PutMapping              // put é o nome da função
+	public ResponseEntity<Tema> put (@Valid @RequestBody Tema tema){
 		return ResponseEntity.ok(repository.save(tema));
+		
 	}
 
+	@ResponseStatus(HttpStatus.NO_CONTENT)
 	@DeleteMapping("/{id}")
 	public void delete(@PathVariable Long id) {
+		Optional <Tema> tema = repository.findById(id); // opção: se existir, boa. Senão, boa tb.
+		if(tema.isEmpty())
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND);
 		repository.deleteById(id);
 	}
